Arrow highlight props | render arrow highlight argument changes
