+++ MPChartLib/src/main/java/com/github/mikephil/charting/data/BarLineScatterCandleBubbleDataSet.java
@@ -4,6 +4,7 @@ package com.github.mikephil.charting.data;
 import android.graphics.Color;
 
 import com.github.mikephil.charting.interfaces.datasets.IBarLineScatterCandleBubbleDataSet;
+import com.github.mikephil.charting.utils.Utils;
 
 import java.util.List;
 
@@ -20,6 +21,9 @@ public abstract class BarLineScatterCandleBubbleDataSet<T extends Entry>
      * default highlight color
      */
     protected int mHighLightColor = Color.rgb(255, 187, 115);
+
+    /** the width of the highlight indicator lines */
+    protected float mHighlightLineWidth = 0.5f;
 
     public BarLineScatterCandleBubbleDataSet(List<T> yVals, String label) {
         super(yVals, label);
@@ -41,8 +45,22 @@ public abstract class BarLineScatterCandleBubbleDataSet<T extends Entry>
         return mHighLightColor;
     }
 
+    /**
+     * Sets the width of the highlight line in dp.
+     * @param width
+     */
+    public void setHighlightLineWidth(float width) {
+        mHighlightLineWidth = width;
+    }
+
+    @Override
+    public float getHighlightLineWidth() {
+        return mHighlightLineWidth;
+    }
+
     protected void copy(BarLineScatterCandleBubbleDataSet barLineScatterCandleBubbleDataSet) {
         super.copy(barLineScatterCandleBubbleDataSet);
         barLineScatterCandleBubbleDataSet.mHighLightColor = mHighLightColor;
+        barLineScatterCandleBubbleDataSet.mHighlightLineWidth = mHighlightLineWidth;
     }
 }
