+++ MPChartLib/src/main/java/com/github/mikephil/charting/interfaces/datasets/IBarLineScatterCandleBubbleDataSet.java
@@ -13,4 +13,10 @@ public interface IBarLineScatterCandleBubbleDataSet<T extends Entry> extends IDa
      * @return
      */
     int getHighLightColor();
+
+    /**
+     * Returns the line-width in which highlight lines are to be drawn.
+     * @return
+     */
+    float getHighlightLineWidth();
 }
