+++ MPChartLib/src/main/java/com/github/mikephil/charting/interfaces/datasets/ILineScatterCandleRadarDataSet.java
@@ -22,12 +22,6 @@ public interface ILineScatterCandleRadarDataSet<T extends Entry> extends IBarLin
     boolean isHorizontalHighlightIndicatorEnabled();
 
     /**
-     * Returns the line-width in which highlight lines are to be drawn.
-     * @return
-     */
-    float getHighlightLineWidth();
-
-    /**
      * Returns the DashPathEffect that is used for highlighting.
      * @return
      */
